Move bookmark API endpoints to their own object
